+++ core-processor/src/main/java/jeff/core/aop/ControllerExceptionResAspect.java
@@ -3,7 +3,7 @@ package jeff.core.aop;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jeff.common.consts.ResponseCode;
-import jeff.common.entity.dto.receive.ResponseObjectFromInnerSystem;
+import jeff.common.entity.dto.inner.InnerCommunicationDto;
 import jeff.common.util.LogUtil;
 import jeff.common.exception.BusyException;
 import jeff.common.exception.MyNotFoundException;
@@ -11,7 +11,6 @@ import jeff.core.exception.OrderException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -22,7 +21,7 @@ import javax.annotation.PostConstruct;
  * 切面程式，攔截控制器所拋出的例外，根據例外去回應給客戶端不同的訊息。
  *
  * <p><del>遭遇錯誤，統一也回Http Code200，會有自訂義的code去表示請求的狀態。</del></p> <---如果這樣做，那麼上游Server的resilience4j不會把錯誤的請求算到斷路器count裡面。
- * 遭遇錯誤時，回應Http狀態碼為非2XX。
+ * 遭遇非預期錯誤時，回應Http狀態碼為非2XX，預期內錯誤則回200。兩者差別在於會不會觸發上游Server的斷路器。
  */
 @Slf4j
 @RestControllerAdvice(basePackages = {"jeff.core.controller"})
@@ -49,24 +48,21 @@ public class ControllerExceptionResAspect {
 
 
     @ExceptionHandler(value = OrderException.class)
-    public ResponseEntity<ResponseObjectFromInnerSystem> handleOrderException(OrderException oe) {
-        return ResponseEntity
-                .status(ResponseCode.Failure.getCode())
-                .body(new ResponseObjectFromInnerSystem(ResponseCode.Failure.getCode(), EMPTY_CONTENT, oe.getMessage()));
+    @ResponseStatus(code = HttpStatus.OK)
+    public InnerCommunicationDto handleOrderException(OrderException oe) {
+        return new InnerCommunicationDto(ResponseCode.Failure.getCode(), EMPTY_CONTENT, oe.getMessage());
     }
 
     @ExceptionHandler(value = MyNotFoundException.class)
-    public ResponseEntity<ResponseObjectFromInnerSystem> handleMyNotFoundException(MyNotFoundException mnfe) {
-        return ResponseEntity
-                .status(ResponseCode.NotFound.getCode())
-                .body(new ResponseObjectFromInnerSystem(ResponseCode.NotFound.getCode(), EMPTY_CONTENT, mnfe.getMessage()));
+    @ResponseStatus(code = HttpStatus.OK)
+    public InnerCommunicationDto handleMyNotFoundException(MyNotFoundException mnfe) {
+        return new InnerCommunicationDto(ResponseCode.NotFound.getCode(), EMPTY_CONTENT, mnfe.getMessage());
     }
 
     @ExceptionHandler(value = BusyException.class)
-    public ResponseEntity<ResponseObjectFromInnerSystem> handleBusyException(BusyException be) {
-        return ResponseEntity
-                .status(ResponseCode.TooManyReq.getCode())
-                .body(new ResponseObjectFromInnerSystem(ResponseCode.TooManyReq.getCode(), EMPTY_CONTENT, be.getMessage()));
+    @ResponseStatus(code = HttpStatus.TOO_MANY_REQUESTS)
+    public InnerCommunicationDto handleBusyException(BusyException be) {
+        return new InnerCommunicationDto(ResponseCode.TooManyReq.getCode(), EMPTY_CONTENT, be.getMessage());
     }
 
     /**
@@ -74,7 +70,7 @@ public class ControllerExceptionResAspect {
      */
     @ExceptionHandler(value = Exception.class)
     @ResponseStatus(code = HttpStatus.INTERNAL_SERVER_ERROR)
-    public ResponseObjectFromInnerSystem handleException(Exception e) {
+    public InnerCommunicationDto handleException(Exception e) {
         logUtil.logError(
                 log,
                 logUtil.composeLogPrefixForSystem(),
@@ -82,7 +78,7 @@ public class ControllerExceptionResAspect {
                 e
         );
 
-        return new ResponseObjectFromInnerSystem(ResponseCode.Failure.getCode(), EMPTY_CONTENT, "Some errors occurred while processing request, please call the application owner.");
+        return new InnerCommunicationDto(ResponseCode.Failure.getCode(), EMPTY_CONTENT, "Some errors occurred while processing request, please call the application owner.");
     }
 
 }