+++ core-processor/src/main/java/jeff/core/filter/ReqContextGenerationFilter.java
@@ -2,6 +2,7 @@ package jeff.core.filter;
 
 import jeff.common.entity.bo.MyRequestContext;
 import jeff.core.entity.bo.MyContentCachingReqWrapper;
+import jeff.core.filter.consts.FilterOrderNumberConst;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;