+++ high-concurrency-processor/src/main/java/jeff/highconcurrency/http/feign/client/CoreProcessorFeignClient.java
@@ -1,7 +1,7 @@
 package jeff.highconcurrency.http.feign.client;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import jeff.common.entity.dto.receive.ResponseObjectFromInnerSystem;
+import jeff.common.entity.dto.inner.InnerCommunicationDto;
 import jeff.highconcurrency.http.feign.config.ReactiveFeignConfigForCoreProcessor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
@@ -17,12 +17,11 @@ import reactor.core.publisher.Mono;
 @Component
 @ReactiveFeignClient(
         name = "core-processor", //如果開啟負載均衡模式，那麼這個名字要和你要打過去的服務集群名稱一樣(或者說要跟該集群註冊進服務治理中心的名稱一樣)
-        path = "/inner",
         configuration = ReactiveFeignConfigForCoreProcessor.class
 )
 public interface CoreProcessorFeignClient {
 
     @PostMapping(value = "/flash-sale-event/query")
-    Mono<ResponseEntity<Mono<ResponseObjectFromInnerSystem>>> getFlashSaleEventInfo(@RequestHeader(value = "myUUID") String myUUID, @RequestBody JsonNode param);
+    Mono<ResponseEntity<Mono<InnerCommunicationDto>>> getFlashSaleEventInfo(@RequestHeader(value = "myUUID") String myUUID, @RequestBody JsonNode param);
 
 }